+++ CHANGELOG.md
@@ -39,6 +39,8 @@
   - GADT's can now span multiple lines ([#102](https://github.com/JustusAdam/language-haskell/issues/102))
   - Implemented existential types ([#39](https://github.com/JustusAdam/language-haskell/issues/39))
   - `data` and `newtype`s can now include kind signatures ([#35](https://github.com/JustusAdam/language-haskell/issues/35))
+- Support for foreign imports (see https://www.haskell.org/onlinereport/haskell2010/haskellch8.html)
+  ([#44](https://github.com/JustusAdam/language-haskell/issues/44))
 - As a substantial internal change the grammar migratd to the YAML format, to make it easier to
   maintain and develop with.
 - Refined many of the tm scopes with new information available from the parser